requirements.txt
agno
openai
python-dotenv
duckduckgo-search
streamlit 
pandas
openai 
duckduckgo-search 
fastapi[standard]
PyGithub
gradio
